requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
phue>=1.1
pyyaml>=6.0
numpy>=1.24.0
